Refactor FormIntermediateStepPage to directly use data.steps in StepVisualization and move step tree construction to StepVisualization component
