+++ packages/nextjs/src/components/pages/form-intermediate-step-page/FormIntermediateStepPage.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Button, BodyLong } from '@navikt/ds-react';
 import { translator } from 'translations';
-import { SelectableStep, StepBase } from 'types/content-props/form-intermediate-step';
+import { StepBase } from 'types/content-props/form-intermediate-step';
 import { ParsedHtml } from 'components/_common/parsedHtml/ParsedHtml';
 import { LenkeBase } from 'components/_common/lenke/lenkeBase/LenkeBase';
 import { useFormIntermediateStepPage } from 'components/pages/form-intermediate-step-page/useFormIntermediateStepPage';
@@ -11,10 +11,7 @@ import { IllustrationStatic } from 'components/_common/illustration/static/Illus
 import { ContentCommonProps, ContentType } from 'types/content-props/_content-common';
 import { PictogramsProps } from 'types/content-props/pictograms';
 import { Taxonomy } from 'types/taxonomies';
-import {
-    StepVisualization,
-    type TreeNode,
-} from 'components/pages/form-intermediate-step-page/step-visualization/StepVisualization';
+import { StepVisualization } from 'components/pages/form-intermediate-step-page/step-visualization/StepVisualization';
 
 import style from './FormIntermediateStepPage.module.scss';
 
@@ -29,16 +26,6 @@ export type FormIntermediateStepPageProps = ContentCommonProps & {
     } & StepBase;
 };
 
-const buildStepTree = (steps: SelectableStep[]): TreeNode[] => {
-    return steps.map((step) => {
-        const node: TreeNode = { label: step.label };
-        if (step.nextStep?._selected === 'next' && step.nextStep.next?.steps) {
-            node.children = buildStepTree(step.nextStep.next.steps);
-        }
-        return node;
-    });
-};
-
 export const FormIntermediateStepPage = (props: FormIntermediateStepPageProps) => {
     const { language, data, displayName } = props;
     const { illustration, formNumbers } = data;
@@ -47,8 +34,6 @@ export const FormIntermediateStepPage = (props: FormIntermediateStepPageProps) =
     const currentStepTitle = currentStepData.title ?? displayName;
 
     const getTranslations = translator('form', language);
-
-    const stepTree = buildStepTree(data.steps);
 
     return (
         <>
@@ -98,7 +83,7 @@ export const FormIntermediateStepPage = (props: FormIntermediateStepPageProps) =
                     </div>
                 )}
             </div>
-            {props.editorView === 'edit' && <StepVisualization steps={stepTree} />}
+            {props.editorView === 'edit' && <StepVisualization steps={data.steps} />}
         </>
     );
 };
