+++ packages/nextjs/src/components/pages/form-intermediate-step-page/step-visualization/StepVisualization.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { SelectableStep } from 'types/content-props/form-intermediate-step';
 import style from './StepVisualization.module.scss';
 
 export type TreeNode = {
@@ -7,7 +8,17 @@ export type TreeNode = {
 };
 
 type Props = {
-    steps: TreeNode[];
+    steps: SelectableStep[];
+};
+
+const buildStepTree = (steps: SelectableStep[]): TreeNode[] => {
+    return steps.map((step) => {
+        const node: TreeNode = { label: step.label };
+        if (step.nextStep?._selected === 'next' && step.nextStep.next?.steps) {
+            node.children = buildStepTree(step.nextStep.next.steps);
+        }
+        return node;
+    });
 };
 
 const TreeItem = ({ node }: { node: TreeNode }) => (
@@ -24,9 +35,11 @@ const TreeItem = ({ node }: { node: TreeNode }) => (
 );
 
 export const StepVisualization = ({ steps }: Props) => {
+    const treeData = buildStepTree(steps);
+
     return (
         <ul className={style.tree}>
-            {steps.map((step, index) => (
+            {treeData.map((step, index) => (
                 <TreeItem key={index} node={step} />
             ))}
         </ul>
